Commit always
code still dose the same but for variation and to
be a reference for me i changed the black function
instead of inserting an anonymous function i made
a function that changes the background color
and passed it as a reference.

The overall behavior remains exactly the same,
but (as learning reference for me) the black
event handler calls a named
function instead of an anonymous one.